Do not use UNAME_SYSNAME check to distinguish between target platforms. Since these variables are defined at Cython compile time, the resulting .c file will compile only on the platform on which Cython was run.
Unfortunately, Cython does not give us access to the C preprocessor.

Therefore, we introduce our own TARGET_PLATFORM variable and run
Cython once for each platform, generating different .c files for each
case. The preprocessor is then used to insert the correct .c file
at C compilation time.

Unfortunately this means that we have to either add a lot of
complexity to the Cython compilation code, or simply hardcode the
files that we want to compile. For now, we've gone with the
latter option.